change udma address to uintptr

+++ src/drv/include/udma.h
@@ -59,8 +59,8 @@ class udma_c
 
         static udma_c& getInstance(void);      
 
-        uint64_t getBufferAddress(uint32_t bufferIndex);        
-        uint64_t getBufferPhysicalAddress(uint32_t bufferIndex);
+        uintptr_t getBufferAddress(uint32_t bufferIndex);        
+        uintptr_t getBufferPhysicalAddress(uint32_t bufferIndex);
         uint32_t getBufferSize(uint32_t bufferIndex);
 
         void dumpUdmaBufferInformation(void);
